// frontend/src/types/patient.ts
export interface Patient {
  case_id: string;
  estimated_body_weight: number;
  age: number;
  bmi: number;
  size_cm: number;
  gcs_summary: string;
  female: number;
  male: number;
  nonbinary: number;
  nationality: string;
  patient_category_intensive_care: number;
  patient_category_high_risk: number;
  patient_category_no_vital_danger: number;
  blood_pressure_id: number;
  body_sensitivity_id: number;
  cardio_id: number;
  condition_basic_id: number;
  condition_consciousness_id: number;
  condition_mental_id: number;
  condition_pre_id: number;
  condition_vitals_id: number;
  death_id: number;
  expert_comment_id: number;
  illness_assessment_id: number;
  injury_id: number;
  intervention_id: number;
  neurology_id: number;
  respiratory_id: number;
  treatment_id: number;
  infections: number;
  lung_disease: number;
  cardiovascular_disease: number;
  abdominal_disease: number;
  metabolic_disease: number;
  other_diseases: number;
  cns_disease: number;
  respiratory_disease: number;
  psychiatric_disease: number;
  gyn_obstetrics_emergencies: number;
}