// frontend/src/types/diagnoses.ts
export interface Diagnoses {
  infections: number;
  lung_disease: number;
  cardiovascular_disease: number;
  abdominal_disease: number;
  metabolic_disease: number;
  other_diseases: number;
  cns_disease: number;
  respiratory_disease: number;
  psychiatric_disease: number;
  gyn_obstetrics_emergencies: number;
}